Adds Promise.lua to flatten out u.system and async.lua behavior
Regardless of whether we're running a job or some async computation, we
should have a unified interface for waiting on the completion of these
processes and computations.

Promises are used to as a location where computation results will arrive
sometime in the future. We are then able to wait over multiple promises
in the same way regardless of what the concurrency mechnism actually is.

The way this is meant to be used is more like how u.system_async used to
work. We start mulitple processes, wait for all results to come in, and
proceed with the rest of the program in a linear fashion.